this is a large change. now all playlist entry data is only loaded by demand. ( when it should be displayed or edited). data is cached internally. the playlist of musikplayer is now not a copy but it is shared with the playlistctrl, if it is being displayed, therefore changes are always synchronized.

+++ trunk/wxMusik/src/Classes/PlaylistInfoCtrl.cpp
@@ -115,7 +115,7 @@ void CPlaylistInfoCtrl::Update()
 	{
 		stTotalVal->SetLabel( wxString::Format( wxT( "%d" ), wxGetApp().Library.GetSongCount() ) );
 		stPlaylistVal->SetLabel( wxString::Format( wxT( "%d" ), m_pIPlaylistInfo->GetCount() ) );
-		int nPlaylistTime = m_pIPlaylistInfo->GetTotalPlayingTimeInSeconds();
+		int nPlaylistTime = (int)m_pIPlaylistInfo->GetTotalPlayingTimeInSeconds();
 		wxTimeSpan PlayTime( 0, 0, nPlaylistTime );
 		stRuntimeVal->SetLabel ( PlayTime.Format() );
 		wxString strsize( wxT("0.0 mb") );
